requirements.txt
python-dotenv
playwright
easyocr
logging
flask
pymongo
numpy
opencv-python-headless